+++ src/views/@common/hooks/api.ts
@@ -1,6 +1,4 @@
-import axios, { AxiosInstance, isAxiosError } from 'axios';
-
-import removeToken from '../utils/removeToken';
+import axios, { AxiosInstance } from 'axios';
 
 const api: AxiosInstance = axios.create({
   baseURL: import.meta.env.VITE_BASE_URL,
@@ -24,48 +22,5 @@ export const getRefreshToken = () => {
 export const setRefreshToken = (token: string) => {
   sessionStorage.setItem(REFRESH_TOKEN, token);
 };
-
-api.interceptors.request.use((config) => {
-  const accessToken = getToken();
-  if (accessToken) {
-    config.headers['Authorization'] = `Bearer ${accessToken}`;
-  }
-  return config;
-});
-
-// 토큰 만료 처리 로직
-const postRefresh = async () => {
-  const accessToken = getToken();
-  const refreshToken = getRefreshToken();
-
-  try {
-    const response = await api.post('/auth/refresh', {
-      accessToken: accessToken,
-      refreshToken: refreshToken,
-    });
-    const newToken = response.data.data;
-    setToken(newToken.accessToken);
-    setRefreshToken(newToken.refreshToken);
-  } catch (err) {
-    // refresh api에서 발생한 에러 처리
-    if (isAxiosError(err) && err.response?.status === 401) {
-      console.log(err);
-      removeToken();
-    }
-  }
-};
-
-api.interceptors.response.use(
-  (res) => res,
-  async (err) => {
-    if (err.response.status === 401) {
-      await postRefresh();
-      const token = getToken();
-      err.config.headers['Authorization'] = `Bearer ${token}`;
-      return api(err.config);
-    }
-    return Promise.reject(err);
-  },
-);
 
 export default api;
